requirements.txt
asgiref==3.9.1
attrs==25.3.0
boto3==1.40.20
botocore==1.40.20
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
oauthlib==3.3.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
